Broke out useValidateAsSetter to a separate file
